split up levels and platform stuff and monsters from the scroller file
